allow controller return value to be empty
Controller methods that are used as route handlers aren't required to
return anything, so we have to make sure we have something we can call
`->can` on before we try to do so.

This means that we may end up ending the OpenTelemetry span before the
request is actually handled: If the handler uses a promise to go async,
but then does not return that promise, the span won't know about it.
We should probably document this, and/or provide a way to delay ending
the span (I'm not sure the otel semantics around WebSockets, but those
don't end after the "response" either).