change how mojolicious request cycle is handled
`around_dispatch` gives us a bit more control over things:

* It works before the static server, so responding with
  static content will still be recorded.
* It wraps all of the actions that are called by the request,
  including intermediate actions.

Then, instead of trying to figure out when the handler is finished,
which can be difficult when async is involved, we add an event handler
to the transaction so we know exactly when it is finished.

This changed how the span is built into two stages: One for the basic
HTTP request and another for the Mojolicious route information. That
change required a rewrite to how the tests worked: Now instead of
testing specific calls, we test the end result of the span.